multi-modal-demo/requirements.txt
timm
pymilvus==2.4.0
openai-clip
pillow
numpy
matplotlib
